Implemented a text parsing optimization.
Amounts, costs and assigned amounts are now parsed as regular amounts.  To
have a full value expression at any of those positions, surround it with
parentheses.  The reason for this is that the amount parser is far simpler and
quicker -- and by far the common case -- compared to the full expression
parser.