Added note to docs about "primary" commodities
Whenever a commodity is exchanged for another in a posting, one of the
two is considered "primary", and the other secondary.  Primariness of a
commodity is remembered, since the --market option only renders balances
into secondary commodities, never primaries.  To render primaries, use
the --exchange=COMMODITY option.

In all of the following examples, the P commodity is considered primary
and the S is secondary (the P at the beginning of the line indicates a
price-setting directive):

    2009/01/01 Sample 1a
        Assets:Brokerage:Stocks                100 S
        Assets:Brokerage:Cash                 -100 P

    P 2009/01/15 00:00:00 S 2 P

    2009/02/01 Sample 2a
        Assets:Brokerage:Stocks                100 S @ 1 P
        Assets:Brokerage:Cash

    P 2009/02/01 00:00:00 S 4 P

    2009/03/01 Sample 3a
        Assets:Brokerage:Stocks                100 S @@ 100 P
        Assets:Brokerage:Cash

    P 2009/03/01 00:00:00 S 8 P

    2009/04/01 Sample 4a
        Assets:Brokerage:Cash                  100 P
        Assets:Brokerage:Stocks               -100 S {1 P}

    P 2009/04/01 00:00:00 S 16 P