Added support for tag/pop parsing directives
There are now "tag/pop" directives, to apply metadata to a range of
transactions (and their postings).  For example, if you wanted a
conceptual "page" of transactions relating to business trip to Chicago,
you could do this:

    tag Location: Chicago
    tag Purpose: Business

    ... transactions go here

    pop
    pop

It would be as if you'd applied "; Location: Chicago", etc., to every
transaction.