A simple, yet significiant, optimization
Every annotated commodity is based on a "referent", or the unannotated
version of that commodity.  When stripping all annotations away,
however, rather than simply returning the referent we were actually
searching for it by name.  There was no reason to do this, not to
mention it was taking up to 7% of the total runtime of some reports.