Fixes bug 565, Transactions with leading whitespace not detected.
This commit only effects src/textual.cc

In textual.cc, the instance_t::read_next_directive() function is used
to break down the input file into major parts. It triggers on the
first char in a line, then calls the appropriate handler for that
particular line type.

This commit adds stricter error checking.  Previously
read_next_directive() only took action if it understood the first
character and silently ignored everything else.  This version rejects
unknown lines and throws appropriate errors.

Any line that starts with whitespace is checked for content, if it has
no content then it is ignored.  If it has content a parse error is
thrown and the user is notified.

This commit also specifically allows % as a single line comment
character.