Fixes bug 609, ledger doesn't handle relative paths from a pipe
When session.cc is reading from /dev/stdin it passes /dev/stidin as
the base path for the parser to work with.  This is inapproporiate as
it is not the base of a file system.  Passing "" as the base restores
sanity.  Although I am not sure why anyone would do this.  I was simply
investigating something I found in the source code.