FAQ entry on how build your own Boost and/or CMake for use with Ledger.
Debian squeeze, which is currently the stable distribution at the time of
this commit, has both a Boost and a CMake that is too old for Ledger.
This FAQ entry explains how to build your own Boost and CMake for use with
Ledger, and the exact commands to type to build and install each, and then
configure, build and install Ledger against those new versions.