Receipt/Invoice are additional information thus better if they're final columns.
Change the column of Receipt and Invoice in the CSV file first, then the
generated ODS file must have the same change propagated, which requires
changes to the column numbers hard-coding in csv2ods.py.

Perhaps if/when this application is refactored these things shouldn't be
hard-coded in this way in the first place.