Make reference and link to Bug Report #789 in fixed directive documentation.
Due to weirdness that's currently true with the existing next branch of
ledger, I believe it's important to tell users in the documentation that
there are some discrepancies in the 'fixed' directive behavior.

The documentation from my previous commit is written to explain what
'fixed' *should* do; adding the bug report link here is a placeholder to
tell users that it may not do what they think it does.

Obviously, if someone closes #789, they should remove this paragraph added
herein.  But, if the bug report is closed, but the documentation lags
behind, the worst that happens is some users have to click through to see
the bug is closed.