fix test runner if filename is specified
If the command line switch "-f" is used the test runner will
not pass the file name of the test file to ledger.
The line "test -f /dev/null" as used in test/baseline/cmd-convert.test
did not work as intended before this.
Also using "-f - " did not work as "$ledger" was noti
prepended to the command.