Support a broader set of possible tags to be placed into the spreadsheet.
I've now made a hard-coded list of potential tags that are supported and
will be linked to in the general ledger spreadsheet.  This list should
probably be in a configuration file of some sort eventually, rather than
hard coded.

Indeed, note that the hard-coding goes into two different scripts, and
thus the lists could easily get out of sync.