Correct regular expression for account queries to avoid inclusion of sub-accounts.
The previous queries had a bug whereby an account in the form "A:B" would
include all transactions for sub accounts such as "A:B:C".

That's not the intended effect.  Entries should appear in the lowest level
account, and not in their parent.

The regular expression now is anchored at start and finish in both queries to
ensure this works correctly.