Bug 893 Ledger reconcile loses alignment
An earlier change to multi-file support stored the actual markers to the beginnings of the transaction/postings.
When reconcile would insert characters it would invalidate those marker and after many items and been
cleared could result in severe misalignment.  This change brings back storing the line-numbers as reported by emacs.