Use "ledger accounts" command rather than the '-F "%150A\n" reg' report solution.
Ledger has had (probably for some time) an "accounts" command that will list
all the accounts from all transactions meeting the other criteria set on the
command line.  That's really what we're looking for here when we build this
chart of accounts, and thus that should be used.

Note that this corrects a subtle bug that wasn't apparent with the old
solution.  With the '-F "%150A\n" reg' solution, accounts that balanced out
to zero for period (e.g., accrual accounts that were emptied during in the
month) did not show up on the reports.  This bug that I didn't know I had
here is thus now fixed by switching to the "accounts" report.