Remove some obsolete material from the build instructions.
It no longer works to create the "build" directory and then build
separate optimized and debug versions.  As confirmed in IRC with sm
and egh, the former instructions resulted in an error:

  $ mkdir build
  $ ./acprep opt make
  make: *** No targets specified and no makefile found.  Stop.
  acprep: ERROR: Execution failed: make
  $ ./acprep debug make
  make: *** No targets specified and no makefile found.  Stop.
  acprep: ERROR: Execution failed: make

Maybe there's a fix for this, but I don't know what it is, so this
commit just makes the build instructions work for the common case.