[doc] Replace .Nm with .Ic where appropriate
According to mdoc (7) "The `.Nm' macro is used for the document title or
subject name.  It has the peculiarity of remembering the first argument
it was called with, which should always be the subject name of the page."

Therefore it should only be used the print `ledger` in the manpage

For the Ledger commands register, balance, etc The .Ic (internal
command) is much better suited.
[ci skip]