Don't explicitly link libpython on OS X
Use -undefined dynamic_lookup to allow Python symbols to be discovered
when the ledger module is imported instead of at build time. Without
this change, the Python interpreter crashes when ledger is imported from
a python that is different from (but compatible with) the python against
which the module was built.