Periods should be initialized from from/since clause rather than earliest transaction date
Bug 1159

Use interval start date (period from/since) to initialize first period.

This allows the offset of a period start to be specified

-p 'every 12 months from 2000-04-01'

will have periods

yyyy-04-01 to yyyy-03-31
....

If no from/since is specified for the period the first transaction date reported is used to initialize the period as before.

added test case regress/1159.test