Make automated transactions work with assertions
This fixes issue #1127.  In my understanding, PR #552 was meant to fix
this, but was incomplete.

Without this patch, automated transactions are invisible to
assertions.

This patch fixes this by adding a flag to the account to tell it that
there is a new posting, analogous to the behavior of finalize().

I dug up issue #1127 too late to find that this is the same solution
proposed by @tbm.  Although I wrote this independently, credit goes to
Martin Michlmayr (@tbm).