Expose post_t::given_cost over python
This fixes #1655 by making the post_t::given_cost variable accessible
over python.

This allows access to the given cost of a posting.  For example, here
it will be "-2 EUR":

A  -2 XXX {1 EUR} [2018-01-01] @@ 2 EUR

If a per-unit cost is given, the given_cost variable will still
contain the cost of the posting.  For example, here it will be
"-4 EUR":

B  -2 XXX {1 EUR} [2018-01-01] @ 2 EUR