Use build matrix to specify Boost versions for Travis CI
In Travis CI, versions of libraries, etc. to build against are
typically specified in a build matrix.  In addition, currently there
is no way to build against the distribution-provided Boost version.

This change uses a build matrix for BOOST_VERSION and allows that
variable to be empty for building against the distribution-provided
Boost version.