Ignore null deferred postings
All-null transactions (i.e. a transaction where all postings have a null
amount) are discarded during parsing and the `xact` object is free'd.
But if the transaction contains a deferred posting this results in a
use-after-free vulnerability because a reference to the deferred posting
is stored in the account object which is later read when deferred
postings are applied after parsing is finished.

Ignore null deferred postings to prevent this – they should not have any
effect any way.

Thanks to Cory Duplantis for reporting this issue and providing an
initial analysis.

Ref TALOS-2017-0304, CVE-2017-2808
Fixes #1723