Fix possible stack overflow in date parsing routine
It is possible to create a stack overflow by giving a date that is
longer than the buffer that is used during date parsing because the
length of the input string is not checked. The `VERIFY` macro is only
enabled when debug-mode is enabled and the `--verify-memory` argument is
used.

Prevent the issue by always checking the input string length and
discarding dates that does not fit in the buffer as invalid.

This issue has been assigned CVE-2017-12482.

Fixes #1224