Fix possible stack overflow in option parsing routine
It is possible to create a stack overflow by giving an option that is
longer than the buffer that is used during option parsing because the
length of the input string is not checked.

Prevent the issue by always checking the input string length and
discarding options that does not fit in the buffer as invalid.

This issue has been assigned CVE-2017-12481.

Thanks to Gwan Yeong Kim for reporting this issue.

Fixes #1222