Use CMAKE_CXX_COMPILER_ID for conditions based on compiler
CMAKE_CXX_COMPILER is the path to the compiler binary and does not
need to follow a specific pattern.  For example, on Linux with GCC and
without an explicit "-DCMAKE_CXX_COMPILER:PATH=" option,
CMAKE_CXX_COMPILER is "/usr/bin/c++" which does not match "g++".
CMAKE_CXX_COMPILER_ID however will always reliably be "Clang" or
"GNU".