Add Travis CI setup for macOS and homebrew-installed Boost
On macOS, CMake detects the Boost.Python component installed by
homebrew only when named "python27".  Thus this change not only adds a
Travis CI setup for macOS, but also a CMake option to switch the
component name between "python" and "python27".  In addition,
precompiling system.hh does not work with the current setup for Clang,
so another CMake option to disable it is added.

The currently used commands to compile specific versions of Boost do
not produce a result that works out of the box on macOS.  It should be
possible just to mimic homebrew's formula for boost-python
(https://github.com/Homebrew/homebrew-core/blob/master/Formula/boost-python.rb),
but for the moment on macOS this change tests only against Boost
installed by homebrew.