test: Test that real account balance assignments are correct
Previously balance assignments to real accounts took virtual postings
into consideration. This means that the checking account would
erroneously receive a posting of $80, when it should receive $-20
to balance against the real account total. This test also includes
a balance assignment on a virtual account for the sake of completeness.